WT-16464 modularise decode tool. Add page wrapper class. Move methods to component files.
